Precompile C sources into boot.a
The commit moves the C sources into the boot
directory, and adds a Makefile so that we can
quickly rebuild the contents. We link boot.a
into the teensy4-rt.

boot.a is checked into version control. Right
now, we'll have to make sure we regenerate the
library if / when the sources change. We should
have some automated check for this...

Closes #2