Add the Teensy4's USB stack
We compile the USB stack in the
teensy4-usb-sys crate. The BSP crate is
responsible for ensuring that the remaining
function calls are satisfied (things like delay
and yield).

We made some mods to the runtime crate to add
the sections for progmem, flash, and dmabuffers.
We can consider renaming these, if we'd like.
We're keeping source compatibility by passing in
the DMAMEM, FLASHMEM, and PROGMEM preprocessor
symbols in the usb-sys crate's build script.

I made a few minor changes to the C sources and
headers in order to suppress compiler warnings.

Finally, we expose the runtime crate through the
BSP crate. I ensured that all three examples still
run in debug / release modes.