Fix initialization of read-only data
We were never initializing read-only data. The
commit moves the section into the overall data
section. Alternatively, we could explicitly copy
over read-only data into the DTCM.

I tried keeping read-only data in FLASH, but the
teensy_cli_loader was yelling about a bad flash
section... Might re-evaluate this later. The
C implementation has the RO data in DTCM, so
we're up to parity with that.