Enable software input on (SION) for I2C pins
Up to this point, we would observe lost
arbitration errors when trying to perform
any I2C operation. Because the pins were
not configured as explicit inputs, we were
unable to operate as an I2C master.

The commit adds methods to enable and disable
software input on. We call the enable method
when configuring I2C pins. With this in place,
we're able to query for a MPU9250's WHO_AM_I.

Remaining changes are to account for improper
code format and some inconsistent register
operations.