Update pad macros to expose the pad ctrl reg
In the I2C implementation, the configure method
has to acquire the correct pad in order to set
pad settings. Rather than have that spread throughout
the code, we can expose the pad register from an
macro-generate pad() method. It reduces some
of the overhead in defining pad configurations.

We update all pad macros to accept the pad
module and pad register type. This was semi-
automated, thanks to VSCode's multicursor.