Begin UART pin configuration
This approach uses more macros than any of the
other peripherals. I'm considering removing the
IntoDaisy and Daisy traits. We can select the
daisy configuration in a peripheral's Pin trait,
in the configure method. If this works out well
for the UART pins, I might backport the same
pattern for the I2C pins.