Thread the PAC's UART peripherals through the HAL
In a prototyping branch, we showed that the UART
peripherals are not Sync when they maintain a
static reference to the UART register block.
The commit updates the internal API to accept
strongly-typed UART peripheral blocks. These
have an unsafe impl Sync associated with them.

This approach should probably be ported back to
the I2C peripherals...