Add support for 16-bit SPI words; use internal CS
The chip's default chip select lets us perform more reliable SPI
transactions. When using any available GPIO for the chip select,
I was finding that chip select toggles were either optimized out,
or were not happening fast enough. This would cause instability in
the SPI transfer, particularly when issuing multiple words. After
switching to the silicon-controlled CS, I see none of these issues!

To initialize the CS pin, pass the appropriate processor pin to the
SPI object's `chip_select_handle`. This returns a `ChipSelect` object.
The object implements `OutputPin`, but the implementation does nothing.
The thinking is that the dummy handle may be used with drivers that
are designed for explicit CS control. Not sure if this is a good
idea. I'm considering removing the embedded_hal::spi::FullDuplex
implementation since it's technically not conformant when used in this
configuration...

Instead, I added support for 16-bit words in a `FullDuplex`
implementation. This required us to set FRAMESZ in the `send` method.
Since we're required to use `send()` before `read()`, it seems like
that's the best place to make the frame distinction.

Update the example to use 16-bit SPI transactions with the MPU9250.
The device is actually expected to use 16-bit transactions! The docs
say that each SPI operation is two or more bytes. We also see better
behavior when moving to 16-bit SPI operations. Also, we update the
example docs, and move around code to make the important things
more obvious.