Define stack start statically
Previous implemenetation would put the stack in the middle of the
DTCM memory region, then reconfigure the stack after copying over
all instructions and data. The stack was placed at the top of the
DTCM blocks.

The new approach skips the reconfiguration, placing the stack at the
top of the DTCM blocks.