Add a convenient patch crate for `cortex-m-rt`
This adds a `cortex-m-rt-patch` directory that contains a tiny crate
that re-exports `teensy4-rt` in its entirety under the crate name
`cortex-m-rt` to provide a convenient way of patching upstream crates.

The hope is to make workarounds for issues like #62 a little easier
pending solving rust-embedded/cortex-m-rt#164 and the reconciliation of
`cortex-m-rt` and `teensy4-rt`.