Specify a unloadable `.unint` memory region
RTIC uses an .unint section for late resources. When the teensy4-rt
doesn't explicitly place this section, the linker is still putting
it somewhere. This causes the Teensy CLI loader to throw an arbitrary
"error reading intel hex file" error. It looks like there's a third
section in the output HEX file, which isn't expected by the CLI
loader.

In any case, see the next commit for a working example.