Implement SysTick timer as a struct; remove delay
This is a breaking change. It closes #36.

Since there's a "systick" feature, we can more readily isolate the
SYSTICK code behind a conditionally-compiled systick module. We
expose a SysTick timer struct with a simple delay() method to provide
the same functionality. The new struct implements the blocking delay
trait of the embedded-hal.

Tested with 'make download_systick' and observing a blinking LED.
All examples build.