Add CI job to check the template from master
Now that we can quickly use cargo-generate, we're not prohibited from
checking the template in the teensy4-rs project. The commit adds a
check on the master branch that will pull-down the template to make sure
it compiles.

Once we start (semi-)formally releasing these crates, we should
update the template to use a release. That'll probably happen after
this RTIC integration settles down.