Remove BSP peripherals, move pins into t40 module
We're moving towards an API that doesn't require us to maintain the
BSP peripheral struct. We'll tell users to use the imxrt-hal
peripherals. We'll add functions to constrain the IOMUXC pads into
Teensy 4.0, or 4.1, pin structs.

Users may now construct the Teensy's Systick struct. The user
provides the cortex-m- SYST peripheral, and we configure it. We rely
on the fact that there's only one safe SYST instance, so we can
own it.

Finally, we remove the USB struct, there's a usb_init() function
to configure the USB logger. We enforce SYST configuration through
the interface.