Define common, t40, and t41 pins
Many pads map to the same pins across the Teensy 4.0 and 4.1 This
commit moves the common pins into a `common` module. We then import
the common pins in the t40 and t41 modules, while also defining the
pins that are unique. Each t40 and t41 module provides its own `Pins`
type, and users will make their own selections on t40 and t41 pins.