Move USB_OTG1 ISRs into examples
After this commit, users are required to call usb::poll() themselves.
We recommend that they do this by defining the USB_OTG1 interrupt
handler, and calling the poll() method in the implementation. The docs
guide users towards the ISR approach.

The change supports advanced USB usage in RTIC. An RTIC user can now
implement the ISR, and check the poll status to see if there's data
from the host. Without this kind of change, the user would need to
poll the reader on an interval. Users can still do this, but now they
can schedule a "read from USB" task as necessary.