Signal unconfigured USB device in serial_write
If you try to log just after initializing the USB stack, the device
might not be configured. This results in a return of 0 from serial
write, which fails the logging assert that all data makes it way
into a transfer descriptor.

We now signal the "not configured" error through the interface, and
we allow unconfigured errors to pass. The updated USB example was
sufficient to demonstrate the issue, then show that the issue was
resolved.