Enable thin LTO for debug builds
We're relying on LTO to strip SysTick duplicates from RTIC example
builds.  This only happens when we're building examples with
'--all-features', which isn't aligned with the "don't build RTIC with
the BSP's default features" requirement. Thin LTO is fast enough, and
lets the user remove the '--release' flag and have a consistent build.