Allocate the heap in OCRAM2
Before this commit, the heap starting address is in DTCM, right after
.uninit. Observable using

  cargo build --target=thumbv7em-none-eabihf --release --no-default-features --features=rt --example=led
  rust-objdump -t target/thumbv7em-none-eabihf/release/examples/led | rg heap

The approach is OK, and consistent with how other systems might separate
their stack and heap. However, it's inconsistent with the official
Teensy 4 runtime, which allocates the heap in OCRAM2.

This commit moves the heap to OCRAM2. Its maximum possible size is
512KB. In practice there may be DMA buffers allocated in the same
section, so not all of the space is available for the heap. Use
heap_len() to compute the amount of heap (in bytes) for a given program.

To demonstrate an offset due to DMA buffers, rerun the commands above
with an example that uses USB (the driver allocates descriptors into
.dma).