Update the BSP to 0.4, use imxrt-hal 0.5
The BSP still re-exports much of its API from the HAL and the RAL. But,
the BSP also includes a board module to simplify application
development. The board module provides a policy for configuring system
clocks, and it tries to make it easier for users to allocate
peripherals. The rewritten examples demonstrate this module.

Now that we have imxrt-log with USB logging support, we no longer need
the C-based USB logging implementation and bindings. This commit drops
all the C code in this repo. The BSP still provides a USB logger, but
it's written in Rust and plays well with the usb-device ecosystem. Users
are free to use the USB peripheral instances to create their own USB
devices.

Since imxrt-hal now has many hardware examples that run on a Teensy 4, I
may not immediately re-implement all hardware examples within this repo.
I figure it's better to generalize hardware examples across boards when
possible. I'm linking to the imxrt-hal examples for users who want to
try out more peripherals on their hardware. Of course, I'll still help
contributors who think an example exclusively belongs here.

Remove outdated documentation from docs/. There's equivalent english /
Rust documentation throughout the imxrt-rs projects that we depend on.