Add MicroMod pins, BSP support for MicroMod boards
We double-checked the pinout and alternate functions against the
official SDK and the reference manual.

"Pins" represent identifiers for processor pads, and they associate to
what's available on the Teensy 4.0 and 4.1 board. Pins do not represent
the pinout on the M.2 edge connector. For example, what we call pin 19
is broken out to the M.2 connector's bottom edge, pin 14.

This commit has no effect on the BSP's runtime. The MicroMod seems to
have additional flash storage. But, for the purposes of the runtime,
we're still pretending that we have Teensy 4.0 flash capacity.